Code manipulated to become 2 components:    * Multicast_sender    * Multicast_receiver
Components currently don't have a _nice_ test harness, but do actively test
each other. Tested the server with multiple receivers on multiple hosts (Linux
and Mac OS X).

Still a number of steps needed for a generic UDP proxy, but the two components
this provides should be of practical use. (If somoene needs to be both a sender
and reciever, then they can just instantiate both inside a component)


Michael.