Simple Multicast transceiver
Listens for packets in the given multicast group. Any data received is
sent to the receiver's outbox. The logic here is likely to be not quite
ideal. When complete though, this will be preferable over the sender and
receiver components since it models what multicast really is rather than
what people tend to think it is.

WARNING: UNTESTED (initial integration of two _working and tested_
components however)


Michael