Kamaelia.Util.Splitter is a component for copying data received on it's inbox to many outboxes.
Kamaelia.Util.LossyConnector is a component that drops data received if it's outbox
is full otherwise passes it on unmodified.  This can be used to
prevent blocking it's source.

Both these classes and their tests are being moved from the main Axon
tree as they are useful general utilities rather than being core
framework.  Previous Axon.Splitter and Axon.LossyConnector.