Some initial tests of ideas for allowing "blocking calls" to generators inside microprocesses. (Specifically allowing a component's "main" to make a blocking call on another generator.
The specific use case this is useful for is during synchronous communications
with other components. Specifically when asking for a resource from another
component.

These changes allow the following:
   def main(self):
      ID = randint(0,20000)
      print "PRE-WAIT"
      yield WaitComplete(self.tick(ID), reactivate(self))
      print "POST-WAIT"

In this example, it's saying don't give me CPU time until self.tick(ID) has
completed it's work. After this, reactivate(self). In order for this to work,
the Scheduler needed changing to understand this. (I've removed some ancient
debug cruft as well, which can come back at some point if useful) Due
to the fact we need the microprocess which the generator self.tick(ID) is
wrapped in to signal to the scheduler at shutdown the reactivate(self) message,
we needed to modify the microprocess class. Finally there's a small collection
of new messages when talking to the scheduler that were added to simplify
testing.

I suspect that the syntactic sugar used here can be made more intuitive.


Michael.