Change to this splitter to be called PlugSplitter for the moment. The idea is to merge this into Kamaelia.Util.Splitter, so that it can be used as follows for the short term:
from Kamaelia.Util.Splitter import PlugSplitter as Splitter

Part of the rationale here is because we're actually finding that we have
multiple different splitters that can exist that can all be used in different
ways with different benefits, none of which are necessarily better than others
but rather have different usecases.


Michael.