This is a somewhat more substantial example/presentation tool than example 8. This tool is under active development at the moment until it can be considered "complete". What that means is not determined - a presentation tool is about showing things, and as a result unless you know what you want to show, what might be in the tool isn't known.
This is driving the development of some of the more visual aspects of
Kamaelia though. There's a real possibility this will turn into a
distributed presentation tool though. The reason for this is it's relatively
trivial to copy the messages across networks that are being sent, which
would allow an element of remote control.

So, whilst this is useful (it's been extended for each presentation I've
given on Kamaelia), it's likely to continue growing :-)


Michael.