Reworked documentation.
Bugfix to DragHandler - did not comply with the return value requirements
of the PyGameApp event dispatch mechanism (didn't return True/False to
claim/reject the event). Intermediary class method introduced to provide
this and call to detect(...) separated out from initializer in the process.

test_DragHandler updated to reflect these changes.

Changes are backward compatible so other code using this should continue to
behave as before. However it would be sensible to change them to use the new
class method too.



Matt