Selector made into a threaded component to allow it to block.
select() call set to block with a small timeout, so the component can still
respond to requests arriving on inboxes/outboxes

sync() method added to threadedcomponent to support selector component.
Regulates the rate of calls to select(), by forcing a sync with _localmain
microprocess's main loop. Without this, the system goes unstable, clogged
with huge numbers of repeats of the event coming out of selector before a
CSA can service it!!

Matt