UnixPipe modified to pause() properly.
NB: This component correctly terminates when its child process terminates;
but this is only because, at the moment, the selector component is
constantly waking it up with "writing will not block" messages; whether it
wants to write or not!

This is an outstanding issue with selector that needs resolving.

Matt