Macro.py - Channel transcoder, added auto-restarting of transcoding into a new file at programe junctions. EIT and AVI files dumped into /data/encodingXXX/ then moved to /data/finishedXXX/ once that programme is complete.
Still suspected of bugs that cause it to busy-wait and freeze at the start
of a new programme, so there is a temporary hack to force it to crash at
that point. Run Macro.py from within a loop, and it will then start
transcoding the next programme okay.

Selector/UnixPipe/Carousel: Bugfixes / hacks to get Macro.py working


Matt