Lots of diagnostic output added (but currently commented out).
Bugfix: Ensured results from select.select() call are only iterated over if
the call is successful.

Bugfix: Ensured that adding a selectable when already being selected over
has noeffect. This was causing awkward failures when the selectable would
later be removed - it would not be removed completely if it had been added
twice.

Matt