Preparation for external control of a second thing.
Added a collection of inboxes to allow external control. 2 are the defaults.
"angle" is created to replace how we were using "inbox". "position" is new
and unused as yet, but you can see we've described how we're going to use
it.

>     Inboxes = {
>        "inbox": "not used",
>        "control": "ignored",
>        "angle" : "We expect to recieve messages telling us the angle of rotation",
>        "position" : "We expect to receive messages telling us the new position",
>     }


Change over from expecting the angle on the inbox.

55c61
<             while self.dataReady("inbox"):
---
>             while self.dataReady("angle"):
57c63
<                 angle = self.recv("inbox")
---
>                 angle = self.recv("angle")

Change from using a simple pipeline to using a graphline. We now
have to explicitly say what links are created.

110,113c116,122
<     from Kamaelia.Util.PipelineComponent import pipeline
<     pipeline(
<        angleIncrement(),
<        rotatingCube(),
---
>     from Kamaelia.Util.Graphline import Graphline
>     Graphline(
>        ROTATION = angleIncrement(),
>        CUBE = rotatingCube(),
>        linkages = {
>           ("ROTATION", "outbox") : ("CUBE", "angle"),
>        }


Michael.