Really naive, easily breakable whiteboard audio, that probably doesn't work!
Assuming it works (and I've no idea if it will, because I've not been able to
test it yet!) it will only work for one server and one client. More than 2
whiteboards and the audio will break because there is no audio mixing in there
yet.

Nastily done - uses a separate TCP connection for the audio.


Matt