Bugfix to parsing of tables. Misinterpreted the spec:
Turns out it is impossible to know whether you have received the
whole table of schedule data. This component therefore just spits
out each fragment as it receives it, packaged up in the same
dict format as before.

Modified the dict format to make it clearer to determine whether
the data is present-following (now&next information) or schedule
(later this week) data.

Matt