Bugfix: a running scheduler doesn't always wake when new microprocesses are activated:
Specifically, a scheduler with no unpaused microprocesses that is waiting (blocked) for
instructions to wake up microprocesses, does not unblock when a new microprocess is activated.

Somehow this bugfix never made it in, even though it was identified
ages ago. I thought a test had been added to identify this, will rectify.

Matt