Basic thin threading wrapping around a component. Currently does a number of things that are suboptimal, but can be dealt with.
What I *think* this gives me is a thread which is seperate from
the component being run by ThreadWrap, meaning it will operate
completely async from the component above. This *should* mean
that if we add threadsafe Queues for sending messages to the
component on the inside, that this should result in messages
passing safely into the threadwrap component and thence safely
to the wrapped component. (ie essentially what Likefile was
trying to do)

Unlike Likefile this takes the same approach for handling component
shutdown, as shown by the testcases at the bottom.


Michael.