I'm changing this because it might fix a problem when using Handle/background reported to the list [Subject "problems with Handle/background", July 2nd, 2008].
When the scheduler is running in another thread (Axon.background), it has no thread and waitForOne has been called (like it's done in Axon.background), if we activate a component (such as Handle), it internally calls _addThread which clears self.extra. If this is done before the:

running = len(self.threads) + self.extra

but after checking the wakeRequests (which is done almost in the previous line), then "running" will be zero and the scheduler will finish without executing the component. Ensuring that the value of "extra" used is the value of extra before the wakeRequests was checked should fix it.