This is the first version of the testing framework working with Handle instead of LikeFile. It still fails ~20% of times for some reason (I'm still working on it), so it's still subject to change the implementation (but the interface is still the same).
While it says that it's testing test_Multicore it is first testing usual Pipeline (as it's done in:
https://kamaelia.svn.sourceforge.net/svnroot/kamaelia/trunk/Sketches/MPS/GSOC08/MultiPipeTest.py
).

The purpose of this commit is to add the new version of KamTest (which mainly changes the LikeFile -> Handle; old background -> new background), and some simple tests as a sample, of how some existing tests can be automated with it. I'll provide more examples shortly, but I'm still facing problems related to the LikeFile -> Handle migration.