Changed to be relevant for Kamaelia Grey.
Also, it's now changed to auto-detect the OS and make sure we're putting
files in the right places for each of them (assuming Linux or Mac OS X).
For other posix unixes we assume that /etc/init.d is OK, and we also
warn non-posix unixes that things are right.


Michael.