Switched to PEP 342 behaviour as an experiment:
Control inbox is a new class of box that doesn't store messages, but instead passes onto a notification callback
which is then used to queue up the incoming exception to throw in the next time the generator is run.

Notable feature here: components automatically generate producerFinished() messages if they terminate normally.

The implementation as it stands:

 * breaks existing components (making 'control' inbox trigger exceptions needs to be optional behaviour)

 * won't work for WaitComplete

 * won't work for threaded components


Matt