Enabling tracing of all messages. In particular, adding Axon.Component.TraceAllSends = True
Causes all messages sent in a system to be traced (dumped to the console for now)

and adding

Axon.Component.TraceAllRecvs = True

Causes all messages received in the system to be traced. The python id of all data sent is also sent which is printed first, enabling you to see what message is sent by what component to what component.

In addition to the existing linkage message tracing:

Axon.Box.ShowAllTransits = True

This makes it *a lot* simpler to find problems.

Michael.