Adds revamped website infrastructure
This is the markdown files, templates, and css for the new version
of the site. This should enable easier maintenance on the webite
from now on. Also all the markdown files have some metadata regarding
when updated, what template to use, some aspects around status/validity
of page content, and so on - with the aim being to improve the site going
forward