Retagged versions of Axon/Kamaelia based on semver
Updating Kamaelia & Axon to use semantic versioning
ala a number of other projects I manage. Initial version is based on the
highest simplified version of both and the patch is based on the number
of overall actual releases so far, based on version history.

This hasn't bumped the versions for the applications since they can
and probably should be evaluated on a case by case basis